Updated Renderer to be more consistent
Changed the renderer so that now when render_Object is created, it calls the object_renderer, in RenderSys to update the quad array.

The update_buff function in renderer is also responsible for create all the indexes necessary by looping through all active render objects.

The active render object is also a list now, that holds pointers to active render_Objects.