Scene load order JSON (See desc)
Now you can load scenes based on the names array in the SceneOrder json file. Use this to load scenes generically at runtime without having to recompile the engine!

Just add the scene name to the json file and launch it. Then call the index number to load it. By default the scene at index 0 is loaded first. Likewise, you can also type in the scene string name instead of the index.